uses SelectWithModal
