+++ app/javascript/src/BackendApis/index.js
@@ -3,5 +3,4 @@
 // $FlowFixMe: name mapper confused because it has 'Form' in file name
 export * from './components/AddBackendForm'
 export * from './components/BackendSelect'
-export * from './components/BackendSelectModal'
 export * from './components/PathInput'
